Really reverting all changes to result merger

+++ sensei-core/src/main/java/com/senseidb/search/node/ResultMerger.java
@@ -1,5 +1,28 @@
 package com.senseidb.search.node;
 
+import java.io.Serializable;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.Collection;
+import java.util.Collections;
+import java.util.Comparator;
+import java.util.HashMap;
+import java.util.HashSet;
+import java.util.Iterator;
+import java.util.LinkedList;
+import java.util.List;
+import java.util.Map;
+import java.util.Map.Entry;
+import java.util.Set;
+
+import org.apache.log4j.Logger;
+import org.apache.lucene.index.IndexReader;
+import org.apache.lucene.search.ScoreDoc;
+import org.apache.lucene.search.SortField;
+import org.apache.lucene.util.PriorityQueue;
+
+import proj.zoie.api.ZoieIndexReader;
+
 import com.browseengine.bobo.api.BoboIndexReader;
 import com.browseengine.bobo.api.BrowseFacet;
 import com.browseengine.bobo.api.BrowseSelection;
@@ -22,26 +45,6 @@ import com.senseidb.search.req.SenseiHit;
 import com.senseidb.search.req.SenseiRequest;
 import com.senseidb.search.req.SenseiResult;
 import com.senseidb.search.req.mapred.impl.SenseiReduceFunctionWrapper;
-import org.apache.log4j.Logger;
-import org.apache.lucene.index.IndexReader;
-import org.apache.lucene.search.ScoreDoc;
-import org.apache.lucene.search.SortField;
-import proj.zoie.api.ZoieIndexReader;
-
-import java.io.Serializable;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Collection;
-import java.util.Collections;
-import java.util.Comparator;
-import java.util.HashMap;
-import java.util.HashSet;
-import java.util.Iterator;
-import java.util.LinkedList;
-import java.util.List;
-import java.util.Map;
-import java.util.Map.Entry;
-import java.util.Set;
 
 public class ResultMerger
 {
@@ -142,9 +145,9 @@ public class ResultMerger
     {
       this.hit = hit;
       this.queue = maxPerGroup <= 1 ? null :
-                   new DocIDPriorityQueue(MY_DOC_COMPARATOR,
-                                          maxPerGroup,
-                                          0);
+          new DocIDPriorityQueue(MY_DOC_COMPARATOR,
+              maxPerGroup,
+              0);
       this.distinctMap = new Map[distinctLength];
       for (int i = 0; i < distinctLength; ++i)
       {
@@ -646,9 +649,9 @@ public class ResultMerger
     return hitList;
   }
 
-  private static final byte UNKNOWN_GROUP_VALUE_TYPE = 0;
-  private static final byte NORMAL_GROUP_VALUE_TYPE = 1;
-  private static final byte LONG_ARRAY_GROUP_VALUE_TYPE = 2;
+  private static final int UNKNOWN_GROUP_VALUE_TYPE = 0;
+  private static final int NORMAL_GROUP_VALUE_TYPE = 1;
+  private static final int LONG_ARRAY_GROUP_VALUE_TYPE = 2;
 
   public static SenseiResult merge(final SenseiRequest req, Collection<SenseiResult> results, boolean onSearchNode)
   {
@@ -704,7 +707,7 @@ public class ResultMerger
     }
     else
     {
-      byte[] rawGroupValueType = new byte[req.getGroupBy().length];  // 0: unknown, 1: normal, 2: long[]
+      int[] rawGroupValueType = new int[req.getGroupBy().length];  // 0: unknown, 1: normal, 2: long[]
 
       PrimitiveLongArrayWrapper primitiveLongArrayWrapperTmp = new PrimitiveLongArrayWrapper(null);
 
@@ -777,7 +780,7 @@ public class ResultMerger
       {
         Set<Object>[] distinctSets = null;
         Object[] distinctValues = null;
-        byte[] distinctValueType = null;
+        int[] distinctValueType = null;
         if (req.getDistinct() != null && req.getDistinct().length != 0)
         {
           distinctSets = new Set[req.getDistinct().length];
@@ -786,7 +789,7 @@ public class ResultMerger
             distinctSets[i] = new HashSet<Object>(req.getMaxPerGroup());
           }
           distinctValues = new Object[distinctSets.length];
-          distinctValueType = new byte[distinctSets.length];
+          distinctValueType = new int[distinctSets.length];
           primitiveLongArrayWrapperTmp = new PrimitiveLongArrayWrapper(null);
         }
         for (Map<Object, HitWithGroupQueue> map : groupMaps)
@@ -807,11 +810,11 @@ public class ResultMerger
             {
               SenseiHit hit = groupMergedIter.next();
               if (distinctSets == null || !checkDuplicate(hit,
-                                                          distinctSets,
-                                                          req.getDistinct(),
-                                                          distinctValues,
-                                                          distinctValueType,
-                                                          primitiveLongArrayWrapperTmp))
+                  distinctSets,
+                  req.getDistinct(),
+                  distinctValues,
+                  distinctValueType,
+                  primitiveLongArrayWrapperTmp))
               {
                 mergedList.add(hit);
               }
@@ -864,7 +867,7 @@ public class ResultMerger
                                         Set<Object>[] distinctSets,
                                         String[] distinct,
                                         Object[] distinctValues,
-                                        byte[] distinctValueType,
+                                        int[] distinctValueType,
                                         PrimitiveLongArrayWrapper primitiveLongArrayWrapperTmp)
   {
     for (int i = 0; i < distinctValues.length; ++i)
@@ -887,7 +890,7 @@ public class ResultMerger
 
   private static List<SenseiHit> buildHitsListNoGroupAccessibles(SenseiRequest req,
                                                                  int topHits,
-                                                                 byte[] rawGroupValueType,
+                                                                 int[] rawGroupValueType,
                                                                  PrimitiveLongArrayWrapper primitiveLongArrayWrapperTmp,
                                                                  Iterator<SenseiHit> mergedIter,
                                                                  int offsetLeft) {
@@ -946,7 +949,7 @@ public class ResultMerger
                                                Collection<SenseiResult> results,
                                                int topHits,
                                                List<FacetAccessible>[] groupAccessibles,
-                                               byte[] rawGroupValueType,
+                                               int[] rawGroupValueType,
                                                PrimitiveLongArrayWrapper primitiveLongArrayWrapperTmp) {
     List<SenseiHit> hitsList = new ArrayList<SenseiHit>(req.getCount());
 
@@ -1149,7 +1152,7 @@ public class ResultMerger
 
   private static List<SenseiHit> buildHitsListNoSortCollector(SenseiRequest req,
                                                               int topHits,
-                                                              byte[] rawGroupValueType,
+                                                              int[] rawGroupValueType,
                                                               Iterator<SenseiHit> mergedIter,
                                                               int offsetLeft) {
 
@@ -1192,12 +1195,12 @@ public class ResultMerger
     return hitsList;
   }
 
-  private static Object extractRawGroupValue(byte[] rawGroupValueType, int groupPosition,
+  private static Object extractRawGroupValue(int[] rawGroupValueType, int groupPosition,
                                              PrimitiveLongArrayWrapper primitiveLongArrayWrapperTmp, SenseiHit hit) {
     return extractRawGroupValue(rawGroupValueType, groupPosition, primitiveLongArrayWrapperTmp, hit.getRawGroupValue());
   }
 
-  private static Object extractRawGroupValue(byte[] rawGroupValueType, int groupPosition,
+  private static Object extractRawGroupValue(int[] rawGroupValueType, int groupPosition,
                                              PrimitiveLongArrayWrapper primitiveLongArrayWrapperTmp, Object rawGroupValue) {
 
     if (rawGroupValueType[groupPosition] == LONG_ARRAY_GROUP_VALUE_TYPE)
@@ -1255,9 +1258,9 @@ public class ResultMerger
     private final Collection<SenseiResult> results;
     private final boolean hasSortCollector;
     private final SenseiHit[] hits;
-    private final byte[] rawGroupValueType;
+    private final int[] rawGroupValueType;
     private final PrimitiveLongArrayWrapper primitiveLongArrayWrapperTmp;
-    private final byte[] distinctValueType;
+    private final int[] distinctValueType;
     private final PrimitiveLongArrayWrapper distinctPrimitiveLongArrayWrapperTmp;
 
     private int totalDocs;
@@ -1267,7 +1270,7 @@ public class ResultMerger
                                 Collection<SenseiResult> results,
                                 boolean hasSortCollector,
                                 SenseiHit[] hits,
-                                byte[] rawGroupValueType,
+                                int[] rawGroupValueType,
                                 PrimitiveLongArrayWrapper primitiveLongArrayWrapperTmp) {
       this.req = req;
       this.maxPerGroup = req.getMaxPerGroup() <= 1 ? 0 : req.getMaxPerGroup();
@@ -1277,7 +1280,7 @@ public class ResultMerger
       this.hits = hits;
       this.rawGroupValueType = rawGroupValueType;
       this.primitiveLongArrayWrapperTmp = primitiveLongArrayWrapperTmp;
-      this.distinctValueType = new byte[distinctLength];
+      this.distinctValueType = new int[distinctLength];
       this.distinctPrimitiveLongArrayWrapperTmp = new PrimitiveLongArrayWrapper(null);
 
       groupMaps = new Map[req.getGroupBy().length];
@@ -1308,16 +1311,16 @@ public class ResultMerger
         if (rawGroupValueType[hit.getGroupPosition()] == LONG_ARRAY_GROUP_VALUE_TYPE)
         {
           groupMaps[hit.getGroupPosition()].put(new PrimitiveLongArrayWrapper(primitiveLongArrayWrapperTmp.data),
-                                                new HitWithGroupQueue(hit,
-                                                                      maxPerGroup,
-                                                                      distinctLength));
+              new HitWithGroupQueue(hit,
+                  maxPerGroup,
+                  distinctLength));
         }
         else
         {
           groupMaps[hit.getGroupPosition()].put(rawGroupValue,
-                                                new HitWithGroupQueue(hit,
-                                                                      maxPerGroup,
-                                                                      distinctLength));
+              new HitWithGroupQueue(hit,
+                  maxPerGroup,
+                  distinctLength));
         }
       }
 
@@ -1353,9 +1356,9 @@ public class ResultMerger
               for (j=0; j<distinctLength; ++j)
               {
                 distinctDataCaches[j] = (FacetDataCache)currentContext.
-                                                        reader.
-                                                        getFacetHandler(req.getDistinct()[j]).
-                                                        getFacetData(currentContext.reader);
+                    reader.
+                    getFacetHandler(req.getDistinct()[j]).
+                    getFacetData(currentContext.reader);
               }
               break;
             }
@@ -1407,11 +1410,11 @@ public class ResultMerger
                       for (k = 0; k < distinctLength; ++k)
                       {
                         tmpScoreDoc.distinctValues[k] = extractRawGroupValue(distinctValueType,
-                                                                             k,
-                                                                             distinctPrimitiveLongArrayWrapperTmp,
-                                                                             distinctDataCaches[k].valArray.getRawValue(
-                                                                               distinctDataCaches[k].orderArray.get(doc)
-                                                                             ));
+                            k,
+                            distinctPrimitiveLongArrayWrapperTmp,
+                            distinctDataCaches[k].valArray.getRawValue(
+                                distinctDataCaches[k].orderArray.get(doc)
+                            ));
                         if (pre == null)
                           pre = hitWithGroupQueue.distinctMap[k].get(tmpScoreDoc.distinctValues[k]);
                       }
@@ -1471,9 +1474,9 @@ public class ResultMerger
                       for (j=0; j<distinctLength; ++j)
                       {
                         distinctDataCaches[j] = (FacetDataCache)currentContext.
-                                                                reader.
-                                                                getFacetHandler(req.getDistinct()[j]).
-                                                                getFacetData(currentContext.reader);
+                            reader.
+                            getFacetHandler(req.getDistinct()[j]).
+                            getFacetData(currentContext.reader);
                       }
                       break;
                     }
